Use AJAX instead socket.io
A website, which are use the protocol https can't
load an iFrame, if the website in the iFrame use
http instead https.(insecure connection)
So it wasn't possible to integrate this site
into "typo3".

But with an AJAX request, we don't need a server
and socket.This is more easier and you can
integrate this solution in all websites.